backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
jinja2==3.1.4
python-multipart==0.0.9
motor==3.4.0
pymongo==4.6.3
redis==5.0.4
groq==0.9.0
chromadb==0.5.3
numpy==1.26.4
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
fastapi-mail==1.4.1
python-dotenv==1.0.1
httpx==0.27.0
email-validator==2.1.1